Added pubsub system and new API endpoint to allow HTTP containers generate by self the needed configs.

+++ api_server/modules/sites/api.py
@@ -20,6 +20,7 @@ from .errors import SITES_ERRORS
 # API server imports
 from connexion import request, NoContent
 from daspanel_connexion_utils import api_fail
+from redis import Redis
 
 def redirects_get_all(cuid):
     tenant = request.headers['Authorization']
@@ -52,6 +53,9 @@ def redirects_edit_item(cuid, rcuid, bdata):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         site = SiteModel.get(cuid=cuid)
     except:
@@ -83,6 +87,9 @@ def redirects_delete_item(cuid, rcuid):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         site = SiteModel.get(cuid=cuid)
     except:
@@ -107,6 +114,9 @@ def redirects_new_item(cuid, bdata):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         site = SiteModel.get(cuid=cuid)
     except:
@@ -129,6 +139,9 @@ def versions_activate(cuid, vcuid):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         site = SiteModel.get(cuid=cuid)
     except:
@@ -146,18 +159,25 @@ def versions_activate(cuid, vcuid):
     site._last_update = datetime.utcnow()
     site.validate()
     site.save()
+
+    conn = Redis(host='daspanel-redis', port=6379, db=0, password=tenant)
+    conn.publish('{0}:daspanel:sites'.format(tenant), 'daspanel.sites')
+
     return site.to_struct(), 200
 
 def versions_clone(cuid, vcuid):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.fs.drivers.plugin_exist(CONFIG.fs.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.fs.active)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         site = SiteModel.get(cuid=cuid)
     except:
         return api_fail(SITES_ERRORS, 'NOTFOUND', cuid)
-    if (not CONFIG.fs.drivers.plugin_exist(CONFIG.fs.active)):
-        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.fs.active)
 
     versions = site.to_struct()['versions']
     vedit = -1
@@ -187,14 +207,20 @@ def versions_clone(cuid, vcuid):
 
     if fs.exists(newver.directory):
         fs.rmtree(newver.directory)
-
     fs.cptree(versions[vedit]['directory'], newver.directory)
+
+    conn = Redis(host='daspanel-redis', port=6379, db=0, password=tenant)
+    conn.publish('{0}:daspanel:sites'.format(tenant), 'daspanel.sites')
+
     return newver.to_struct(), 201
 
 def versions_edit_item(cuid, vcuid, bdata):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         site = SiteModel.get(cuid=cuid)
     except:
@@ -216,12 +242,21 @@ def versions_edit_item(cuid, vcuid, bdata):
     site._last_update = datetime.utcnow()
     site.validate()
     site.save()
+
+    conn = Redis(host='daspanel-redis', port=6379, db=0, password=tenant)
+    conn.publish('{0}:daspanel:sites'.format(tenant), 'daspanel.sites')
+
     return versions[vedit], 200
 
 def versions_delete_item(cuid, vcuid):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.fs.drivers.plugin_exist(CONFIG.fs.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.fs.active)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         site = SiteModel.get(cuid=cuid)
     except:
@@ -229,8 +264,6 @@ def versions_delete_item(cuid, vcuid):
     if site.active_version == vcuid:
         return api_fail(SITES_ERRORS, 'VERSIONISACTIVE', vcuid, site.sitedescription)
         #return {"errors": [], "message": "Cann't delete current active version of the site"}, 400
-    if (not CONFIG.fs.drivers.plugin_exist(CONFIG.fs.active)):
-        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.fs.active)
 
     #somelist[:] = [tup for tup in somelist if determine(tup)]
     versions = site.to_struct()['versions']
@@ -251,6 +284,10 @@ def versions_delete_item(cuid, vcuid):
     site._last_update = datetime.utcnow()
     site.validate()
     site.save()
+
+    conn = Redis(host='daspanel-redis', port=6379, db=0, password=tenant)
+    conn.publish('{0}:daspanel:sites'.format(tenant), 'daspanel.sites')
+
     return NoContent, 204
 
 def versions_get_item(cuid, vcuid):
@@ -285,12 +322,15 @@ def versions_new_item(cuid, bdata):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.fs.drivers.plugin_exist(CONFIG.fs.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.fs.active)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         site = SiteModel.get(cuid=cuid)
     except:
         return api_fail(SITES_ERRORS, 'NOTFOUND', cuid)
-    if (not CONFIG.fs.drivers.plugin_exist(CONFIG.fs.active)):
-        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.fs.active)
 
     newver = SiteVersion(**bdata)
     newver.date = datetime.utcnow()
@@ -303,6 +343,9 @@ def versions_new_item(cuid, bdata):
 
     fs = CONFIG.fs.drivers.get_instance(CONFIG.fs.active, 'DasFs', tenant=tenant, bucket=tenant)
     fs.mkdir(newver.directory)
+
+    conn = Redis(host='daspanel-redis', port=6379, db=0, password=tenant)
+    conn.publish('{0}:daspanel:sites'.format(tenant), 'daspanel.sites')
 
     return newver.to_struct(), 201
 
@@ -323,62 +366,101 @@ def get_httpconf(hostname):
     sites = SiteModel.all()
     entries = []
     for site in sites:
+        sitecfg = {
+            '_cuid': site._cuid,
+            'sitedescription': site.sitedescription,
+            'enabled': True,
+            'configs': []
+        }
         for version in site.versions:
             vrec = {
-                'siteuuid': site._cuid,
                 'sitetype': version.sitetype,
                 'engine': version.runtime,
                 'name': '{0}.v.{1}.sites.{2}'.format(version._cuid, site._cuid, hostname),
                 'ssl': False,
                 'sslcert': '',
                 'certpath': '',
-                'dir': version.directory
+                'dir': version.directory,
+                'domain': hostname
             }
+            sitecfg['configs'].append(vrec)
             if not site.urlprefix == site._cuid:
-                vrec['name'] += ' ' + '{0}.v.{1}.sites.{2}'.format(version._cuid, site.urlprefix, hostname)
-            entries.append(vrec)
+                vrec = {
+                    'sitetype': version.sitetype,
+                    'engine': version.runtime,
+                    'name': '{0}.v.{1}.sites.{2}'.format(version._cuid, site.urlprefix, hostname),
+                    'ssl': False,
+                    'sslcert': '',
+                    'certpath': '',
+                    'dir': version.directory,
+                    'domain': hostname
+                }
+                sitecfg['configs'].append(vrec)
 
         for host in site.redirects:
             result, host_version = site.get_version(host.version)
             if not result:
                 return api_fail(SITES_ERRORS, 'VERSIONNOTFOUND', host.version)
             hrec = {
-                'siteuuid': site._cuid,
                 'sitetype': host_version['sitetype'],
                 'engine': host_version['runtime'],
-                'name': host.hosturl + '.' + host.domain,
+                'name': '{0}.{1}'.format(host.hosturl, host.domain),
                 'ssl': host.ssl,
                 'sslcert': '',
                 'certpath': '/certs',
-                'dir': host_version['directory']
+                'dir': host_version['directory'],
+                'domain': host.domain
             }
-            if host.hosturl == 'www':
-                hrec['name'] += ' ' + host.domain
             if host.ssl:
                 hrec['certpath'] = '/certs/{0}'.format(host.domain)
-            entries.append(hrec)
+            sitecfg['configs'].append(hrec)
+            if host.hosturl == 'www':
+                hrec = {
+                    'sitetype': host_version['sitetype'],
+                    'engine': host_version['runtime'],
+                    'name': '{0}'.format(host.domain),
+                    'ssl': host.ssl,
+                    'sslcert': '',
+                    'certpath': '/certs',
+                    'dir': host_version['directory'],
+                    'domain': host.domain
+                }
+                if host.ssl:
+                    hrec['certpath'] = '/certs/{0}'.format(host.domain)
+                sitecfg['configs'].append(hrec)
 
         result, cur_version = site.get_version(site.active_version)
         if not result:
             return api_fail(SITES_ERRORS, 'VERSIONNOTFOUND', site.active_version)
-
         master = {
-            'siteuuid': site._cuid,
             'sitetype': cur_version['sitetype'],
             'engine': cur_version['runtime'],
             'name': '{0}.sites.{1}'.format(site._cuid, hostname),
             'ssl': False,
             'sslcert': '',
             'certpath': '',
-            'dir': site.active_dir
+            'dir': site.active_dir,
+            'domain': hostname
         }
+        sitecfg['configs'].append(master)
         if not site.urlprefix == site._cuid:
-            master['name'] += ' ' + '{0}.sites.{1}'.format(site.urlprefix, hostname)
-        entries.append(master)
+            master = {
+                'sitetype': cur_version['sitetype'],
+                'engine': cur_version['runtime'],
+                'name': '{0}.sites.{1}'.format(site.urlprefix, hostname),
+                'ssl': False,
+                'sslcert': '',
+                'certpath': '',
+                'dir': site.active_dir,
+                'domain': hostname
+            }
+            sitecfg['configs'].append(master)
+
+        entries.append(sitecfg)
     return entries
 
+
 def get_all():
-    print(request.headers)
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
@@ -391,6 +473,8 @@ def new_item(bdata):
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
     if (not CONFIG.fs.drivers.plugin_exist(CONFIG.fs.active)):
         return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.fs.active)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
 
     newrec = SiteModel(**bdata)
     newrec.urlprefix = newrec._cuid
@@ -414,16 +498,23 @@ def new_item(bdata):
     if not fs.exists(newver.directory):
         fs.mkdir(newver.directory)
 
+    pubsub = CONFIG.pubsub.drivers.get_instance(CONFIG.pubsub.active, 'DasPubSub', password=tenant)
+    pubsub.publish('{0}:daspanel:sites'.format(tenant), 'daspanel.sites')
+
     return newrec.to_struct(), 201
 
 def edit_item(cuid, bdata):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         rec2edit = SiteModel.get(cuid=cuid)
     except:
         return api_fail(SITES_ERRORS, 'NOTFOUND', cuid)
+
     result, site_id = rec2edit.urlprefix_used(bdata["urlprefix"])
     if result and (not cuid == site_id):
         return api_fail(SITES_ERRORS, 'URLPREFIXEXIST', bdata["urlprefix"], site_id)
@@ -433,22 +524,33 @@ def edit_item(cuid, bdata):
     rec2edit._last_update = datetime.utcnow()
     rec2edit.validate()
     rec2edit.save()
+
+    conn = Redis(host='daspanel-redis', port=6379, db=0, password=tenant)
+    conn.publish('{0}:daspanel:sites'.format(tenant), 'daspanel.sites')
+
     #return {"newpaswd": "xxxxx"}, 200
     
 def delete_item(cuid):
     tenant = request.headers['Authorization']
     if not tenant == os.environ['DASPANEL_SYS_UUID']:
         return api_fail(DASPANEL_ERRORS, 'INVALIDAPIKEY', tenant)
+    if (not CONFIG.fs.drivers.plugin_exist(CONFIG.fs.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.fs.active)
+    if (not CONFIG.pubsub.drivers.plugin_exist(CONFIG.pubsub.active)):
+        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.pubsub.active)
+
     try:
         rec2delete = SiteModel.get(cuid=cuid)
     except:
         return api_fail(SITES_ERRORS, 'NOTFOUND', cuid)
-    if (not CONFIG.fs.drivers.plugin_exist(CONFIG.fs.active)):
-        return api_fail(DASPANEL_ERRORS, 'FSMISSINGDRIVER', CONFIG.fs.active)
 
     fs = CONFIG.fs.drivers.get_instance(CONFIG.fs.active, 'DasFs', tenant=tenant, bucket=tenant)
     if fs.exists('content/' + cuid):
         fs.rmtree('content/' + cuid)
+
+    conn = Redis(host='daspanel-redis', port=6379, db=0, password=tenant)
+    conn.publish('{0}:daspanel:sites'.format(tenant), 'daspanel.sites')
+
     rec2delete.delete()
 
 def chgpwd_item(cuid, bdata):
